Tune movement range
With the v1 firmware, we had to set a Z offset in the menus and added an
X, Y offset at the start of the gcode. With the new firmware, we can set
a home offset for X, Y and Z at once. That makes things easier.

Soft endstops are also enabled, which measure the distance from the home
position to the other end of the movement range. That means that when
you move an axis through the printer menu, it won't go further than
mechanically possible. There is an option to disable the limit, for
example to use a negative Z when configuring the home offsets.