Use logical coordinates for bed leveling mesh
When configuring the mesh, it defaulted to native coordinates, which
does not make sense to me: you print within the range of logical
coordinates. For example, X0 on my printer is outside the heat bed area.
I can't measure Z there. So use the logical coordinate range instead.
Which requires that we re-calculate the mesh on each reset.