Update M206 Z after probing
By the looks of it, ABL itself only stores relative offsets and still
needs a global M206 shift up and down. Since we do measure that shift
during the probe we can also set that immediately.

I've had to disable updating the current position, that somehow did not
work for me and I always had to re-home first when that was in place.