Move preheat settings to advanced settings
In our setup Cura / OrcaSlicer determines the print temperature. So this
setting can be confusing to users.

Yes, it does configure the preheat in the menu, but when that's used
(removing/installing filament) the defaults are good enough. And if
desired, they can still be changed in the advanced settings.