Fix high Z clearance too high when probing
We noticed that since the v2.1.3 upgrade the probe height in between
probets was significantly more. It looks like v2.1.3 has some changes,
like "🧑💻 Probe flag in do_z_clearance" (f53217b2c5) that add the z
offset of the probe when doing z clearance for probing.

This adds some configuration and code to tweak the behaviour to our
liking again. Our probe needs adjustment, but we don't need to include
its distance on top of the offset we configure.